+++ CHANGELOG.md
@@ -27,6 +27,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Removed the `mode` configuration parameter from the `Symlinker` extension.
 - Removed the `StubProphecy` and `FunctionProphecy` classes.
 - Removed the `WPHealthcheck` class in favor of reports provided by the `WordPress\Installation` class.
+- Removed the `tad\WPBrowser\Utils\Map` class
 
 ## [3.1.6] 2022-04-28;